Placed all exceptions in ex2 in a separate module
